Poprawka usuwania

+++ src/main/java/com/example/bikeRenting/service/tech/TechMainService.java
@@ -2,6 +2,10 @@ package com.example.bikeRenting.service.tech;
 
 import com.example.bikeRenting.constants.RoleConstants;
 import com.example.bikeRenting.dto.response.UserDTO;
+import com.example.bikeRenting.exception.EntityNotFoundException;
+import com.example.bikeRenting.model.entity.BikeStation;
+import com.example.bikeRenting.model.entity.User;
+import com.example.bikeRenting.model.entity.UserStatus;
 import com.example.bikeRenting.repository.UserRepository;
 import com.example.bikeRenting.service.mapping.user.UserMappingService;
 import com.example.bikeRenting.service.user.LoginService;
@@ -44,7 +48,16 @@ public class TechMainService implements TechService {
     public  UserDTO deleteTech(long techId) {
         var user = userRepository.findById(techId)
                 .orElseThrow(()-> new RuntimeException( "Tech with id " + techId + " does not exist."));
+
+        if(checkIfDeleted(user)) {
+            throw new EntityNotFoundException("Tech with id " + techId + " does not exist.");
+        }
+
         user.setRoles(null);
         return loginService.deleteUser(techId);
+    }
+
+    private boolean checkIfDeleted(User user) {
+        return UserStatus.DELETED == user.getStatus();
     }
 }
